Copy PhysicsEvent in PhysicsConverter
This converter is dedicated to converting data from simulation only. To
avoid confusion it is better to move PhysicsEvent content into
PhysicsConverter. This one only file will act as simulation data
converter only and cannot be used by others for TB data preparation or
TB data analysis.
